pyproject.toml
[project]
name = "email-handler-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "ipython>=9.12.0",
    "langchain-core>=1.2.26",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.2",
    "structlog>=25.5.0",
]
